requirements.txt
asteroid-filterbanks==0.4.0
cffi==1.17.0
colorama==0.4.6
filelock==3.15.4
fsspec==2024.6.1
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
pillow==10.2.0
pycparser==2.22
scikit-learn==1.5.1
scipy==1.14.0
soundfile==0.12.1
sympy==1.13.2
threadpoolctl==3.5.0
# torch==2.4.0+cu121
# torchaudio==2.4.0+cu121
# torchvision==0.19.0+cu121
torch
torchaudio
torchvision

tqdm==4.66.5
typing_extensions==4.12.2
pycryptodome==3.20.0
simhash
lshashpy3
ssdeep
graycode